src/main/java/com/rovo98/leetcode/l75/day3/P206ReverseLinkedList.java
package com.rovo98.leetcode.l75.day3;

import com.rovo98.leetcode.common.ListNode;

/**
 * https://leetcode.com/problems/reverse-linked-list
 */
public class P206ReverseLinkedList {
    public ListNode reverseList(ListNode head) {
        if (head == null || head.next == null) return head;
        ListNode result = null;
        while (head != null) {
            result = new ListNode(head.val, result);
            head = head.next;
        }
        return result;
    }
}
/*
Definition for singly-linked list.
public class ListNode {
    int val;
    ListNode next;
    ListNode() {}
    ListNode(int val) { this.val = val; }
    ListNode(int val, ListNode next) { this.val = val; this.next = next; }
}
 */
